+++ learn-ucore/user/libs/ulib.h
@@ -21,7 +21,7 @@ void __noreturn __panic(const char *file, int line, const char *fmt, ...);
 
 // static_assert(x) will generate a compile-time error if 'x' is false.
 #define static_assert(x)                                \
-    switch (x) { case 0: case (x): ; }
+    switch ((int)(x)) { case 0: case (x): ; }
 
 int fprintf(int fd, const char *fmt, ...);
